extract fair top k rescoring out of the rescorer context
