+++ src/main/java/com/purbon/search/fair/lib/FairnessTableLookup.java
@@ -1,0 +1,7 @@
+package com.purbon.search.fair.lib;
+
+public interface FairnessTableLookup {
+
+    float fairness(int trials, float proportion, float significance);
+
+}
